requirements.txt
anthropic>=0.39.0
elevenlabs>=1.0.0
google-genai>=0.3.0
openai-whisper>=20231117
python-dotenv>=1.0.0
Pillow>=10.0.0
requests>=2.31.0
